Fix incorrect text at end of chapter 6
Because the monster AI's `join` includes the `Monster` tag, the player is being excluded from this loop whether they have a `Name` or not. It's not relevant either way to the logic of the AI anyway, as it gets the player's position via the `Point` resource created earlier in the chapter.