The first update of MonsterAI can't take RunState as it's not embedded yet.
At this stage of the code RunState is still embedded in State and not an entity in the world so it can't be passed over in data. A few sections down when RunState is expanded the proper code is shown there. Running the code at this point will panic otherwise.